Refactor project into separate modules
