+++ README.md
@@ -1,0 +1,13 @@
+# Aho-Corasick
+
+Este proyecto implementa un buscador de patrones basado en el autómata Aho-Corasick.
+
+## Compilación
+
+Para compilar la aplicación ejecute:
+
+```bash
+g++ -std=c++17 PatternMatcher.cpp ui.cpp main.cpp -o proyecto
+```
+
+Luego, ejecute `./proyecto` para iniciar el menú interactivo.
